test on the function

+++ ChemFlow/src/ScoreFlow_functions.bash
@@ -294,7 +294,8 @@ for LIGAND in ${LIGAND_LIST[@]} ; do
 
     # Check if the charges are in ChemBase.
     if [ "${DONE_CHARGE}" == "false" ] && [ -s ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.lst ] && [ -s ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.mol2 ] ; then
-        if [ $(grep ${LIGAND_NAME} ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.lst) == ${LIGAND_NAME} ] ; then
+        if [ "$(grep ${LIGAND_NAME} ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.lst)" == ${LIGAND_NAME} ] ; then
+            echo "${CHARGE} charges found in ChemBase for ${LIGAND}"
 
             awk -v LIGAND=${LIGAND_NAME} '$0 ~ LIGAND {flag=1;next}/BOND/{flag=0}flag' ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.mol2 | awk '/1 MOL/&&!/TEMP/ {print $9}' > charges.dat
             antechamber -i ligand_gas.mol2 -o ligand_${CHARGE}.mol2 -fi mol2 -fo mol2 -cf charges.dat -c rc -pf yes &> /dev/null
@@ -306,6 +307,7 @@ for LIGAND in ${LIGAND_LIST[@]} ; do
 
     # If it was not in ChemBase, look into the LigFlow folder.
     if [ "${DONE_CHARGE}" == "false" ] && [ -f ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 ] ; then
+        echo "${CHARGE} charges found in LigFlow for ${LIGAND}"
 
         awk '/1 MOL/&&!/TEMP/ {print $9}' ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 > charges.dat
         antechamber -i ligand_gas.mol2 -o ligand_${CHARGE}.mol2 -fi mol2 -fo mol2 -cf charges.dat -c rc -pf yes &> /dev/null
@@ -316,13 +318,15 @@ for LIGAND in ${LIGAND_LIST[@]} ; do
 
     # If it was not in ChemBase or LigFlow, compute them.
     if [ ${DONE_CHARGE} == "false" ] ; then
-        echo "Compute"
-#        echo -ne "Computing ${CHARGE} charges for ${LIGAND}     \r"
+        echo "Computing ${CHARGE} charges for ${LIGAND}"
         case ${CHARGE} in
         "bcc")
             # Compute am1-bcc charges
             if [ ! -f ligand_bcc.mol2 ] ; then
-                echo "cd ${RUNDIR}/${LIGAND} ; antechamber -i ligand_gas.mol2 -fi mol2 -o ligand_bcc.mol2 -fo mol2 -c bcc -s 2 -eq 1 -rn MOL -pf y -dr no &> antechamber.log ; cp ligand_bcc.mol2 ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2" >> ${RUNDIR}/charges.xargs
+                echo "cd ${RUNDIR}/${LIGAND} ; antechamber -i ligand_gas.mol2 -fi mol2 -o ligand_bcc.mol2 -fo mol2 -c bcc -s 2 -eq 1 -rn MOL -pf y -dr no &> antechamber.log" >> ${RUNDIR}/charges.xargs
+#                if [ ! -f ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 ] ; then
+#                    echo "cp ligand_bcc.mol2 ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2" >> ${RUNDIR}/charges.xargs
+#                fi
             fi
             ;;
         "resp")
@@ -335,7 +339,9 @@ for LIGAND in ${LIGAND_LIST[@]} ; do
 
                 # Read Gaussian output and write new optimized ligand with RESP charges
                 antechamber -i ligand.gout -fi gout -o ligand_resp.mol2 -fo mol2 -c resp -s 2 -rn MOL -pf y -dr no &>> ${RUNDIR}/antechamber.log
-                cp ligand_resp.mol2 ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2
+                if [ ! -f ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 ] ; then
+                    cp ligand_resp.mol2 ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2
+                fi
             fi
         ;;
         esac
